Improve Legend
Improve on the legends usability. Instead of being a part of the
SteeringData, it is now a part of a MovingEntity. Every SteeringData can
 then add the legend items that it wants to use. The legend items can
 also be removed to turn off that part of the legend.

Add the DataPrototype property to ISteeringBehavior. It is used to add
all the legend items before any calculations have completed. It is
expected to return an empty SteeringData object.

Add VectorExtensions to add the DrawVector function.